Fix:コメント一覧部分をShowOutputsComments.vueを作成し、component化
